+++ src/DonorProfiles/resources/js/app/components/donation-table/donation-row/index.js
@@ -1,4 +1,5 @@
 import { Link } from 'react-router-dom';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 const { __ } = wp.i18n;
 
 const DonationRow = ( { donation } ) => {
@@ -14,18 +15,27 @@ const DonationRow = ( { donation } ) => {
 				{ form.title }
 			</div>
 			<div className="give-donor-profile-table__column">
-				{ payment.date }
+				<div className="give-donor-profile-table__donation-date">{ payment.date }</div>
+				<div className="give-donor-profile-table__donation-time">{ payment.time }</div>
 			</div>
 			<div className="give-donor-profile-table__column">
 				<div className="give-donor-profile-table__donation-status">
-					{ payment.status }
+					<div className="give-donor-profile-table__donation-status-indicator" style={ { background: payment.status.color } } />
+					<div className="give-donor-profile-table__donation-status-label">
+						{ payment.status.label }
+					</div>
 				</div>
+				{ payment.mode !== 'live' && (
+					<div className="give-donor-profile-table__donation-test-tag">
+						{ __( 'Test Donation', 'give' ) }
+					</div>
+				) }
 			</div>
 			<div className="give-donor-profile-table__pill">
 				<div className="give-donor-profile-table__donation-id">ID: { id }</div>
 				<div className="give-donor-profile-table__donation-receipt">
 					<Link to={ `/donation-history/${ id }` }>
-						{ __( 'View Receipt', 'give' ) }
+						{ __( 'View Receipt', 'give' ) } <FontAwesomeIcon icon="arrow-right" />
 					</Link>
 				</div>
 			</div>